Added drawer header
Uses Pacifico font. The license will be added with the other licenses. You can find the font here with its license: https://fonts.google.com/specimen/Pacifico
The statusBar is quite ugly with its color overlapping the drawer. It will probably change in a couple of commits for a translucent status bar and toolbars that fitSystemWindow